refactor: store 폴더로 묶기
